requirements.txt
Flask==2.2.2
Pillow==9.4.0
Pillow==10.1.0
Pillow==10.1.0
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==2.2.2
